+++ packages/expo-secure-store/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 🎉 New features
+
+- Added `requireAuthentication` and `authenticationPrompt` parameters to `SecureStoreOptions` options object used in `SecureStore.{deleteItemAsync, getItemAsync, setItemAsync}` methods to enable user authentication while accessing Secure Store. ([#14512](https://github.com/expo/expo/pull/14512) by [@j-piasecki](https://github.com/j-piasecki))
+
 ## Unpublished
 
 ### 🛠 Breaking changes